Refactor: PhotoSwipe를 지연 청크로 분리하고 불필요한 클라이언트 경계 제거
- photoswipe 코어는 react-photoswipe-gallery가 정적 import — 소비처가 아니라 컴포넌트 단위로 나눠야 청크가 분리됨
- 서버 컴포넌트의 dynamic()은 코드 스플리팅 미지원(Next.js Lazy Loading 가이드, Codex 검증 지적)이라 'use client' wrapper(PhotoSwipeLazy) 안에서 dynamic 수행 — LocationMapClient 선례와 같은 모양
- ssr 기본값 유지로 주보 이미지는 SSR HTML에 남음 (LCP·SEO 회귀 없음)
- LatestBulletinImages는 훅·핸들러가 없어 'use client' 제거, 서버 컴포넌트로 강등
- dev 실측: 주보 목록·상세 모두 lightbox 열림/닫힘 정상
